Centralize auto_set_and_close_loop fixture
