+++ tests/adapters/broker/redis_broker/test_processing_subscriber_messages.py
@@ -1,5 +1,3 @@
-import asyncio
-
 import pytest
 
 from protean.core.subscriber import BaseSubscriber
@@ -24,20 +22,6 @@ def clear_terms():
 
     global terms
     terms = []
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 @pytest.mark.redis
