+++ tests/server/test_engine_exits.py
@@ -1,24 +1,9 @@
-import asyncio
 import logging
 
 import pytest
 
 from protean.domain import Domain
 from protean.server.engine import Engine
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 @pytest.mark.no_test_domain
