+++ tests/server/test_engine_initialization.py
@@ -1,22 +1,6 @@
 import asyncio
 
-import pytest
-
 from protean import Engine
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 def test_that_engine_can_be_initialized_from_a_domain_object(test_domain):
