+++ tests/server/test_engine_run.py
@@ -34,20 +34,6 @@ class UserEventHandler(BaseEventHandler):
 
 
 @pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
-
-
-@pytest.fixture(autouse=True)
 def register_elements(test_domain):
     test_domain.config["event_processing"] = Processing.ASYNC.value
     test_domain.register(User, stream_category="authentication")
