+++ tests/server/test_error_handling.py
@@ -1,4 +1,3 @@
-import asyncio
 from uuid import uuid4
 
 import pytest
@@ -33,20 +32,6 @@ class UserEventHandler(BaseEventHandler):
     @handle(Registered)
     def send_notification(self, event: Registered) -> None:
         some_function()
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 @pytest.fixture(autouse=True)
