+++ tests/server/test_event_handling.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import asyncio
 from uuid import uuid4
 
 import pytest
@@ -55,20 +54,6 @@ def reset_counter():
     counter = 0
 
     yield
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 @pytest.mark.asyncio
