+++ tests/server/test_gathering_broker_subscriptions.py
@@ -1,5 +1,3 @@
-import asyncio
-
 import pytest
 
 from protean.core.subscriber import BaseSubscriber
@@ -10,20 +8,6 @@ from protean.utils import fqn
 class DummySubscriber(BaseSubscriber):
     def __call__(self, data: dict):
         pass
-
-
-@pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
 
 
 @pytest.fixture(autouse=True)
