+++ tests/server/test_processing_broker_messages.py
@@ -19,20 +19,6 @@ class DummySubscriber(BaseSubscriber):
 
 
 @pytest.fixture(autouse=True)
-def auto_set_and_close_loop():
-    # Create and set a new loop
-    loop = asyncio.new_event_loop()
-    asyncio.set_event_loop(loop)
-
-    yield
-
-    # Close the loop after the test
-    if not loop.is_closed():
-        loop.close()
-    asyncio.set_event_loop(None)  # Explicitly unset the loop
-
-
-@pytest.fixture(autouse=True)
 def clear_terms():
     yield
